64-bit support

+++ source/d_edge.c
@@ -195,7 +195,7 @@ void D_DrawSurfaces (void)
 			d_zistepv = s->d_zistepv;
 			d_ziorigin = s->d_ziorigin;
 
-			D_DrawSolidSurface (s, (int)s->data & 0xFF);
+			D_DrawSolidSurface (s, (intptr_t)s->data & 0xFF);
 			D_DrawZSpans (s->spans);
 		}
 	}
